Rename `separate-draft-pr-gist-creation-button` to `one-click-pr-or-gist`
