+++ source/refined-github.ts
@@ -147,7 +147,7 @@ import './features/reload-failed-proxied-images';
 import './features/clean-rich-text-editor';
 import './features/highlight-collaborators-and-own-conversations';
 import './features/embed-gist-via-iframe';
-import './features/separate-draft-pr-gist-creation-button';
+import './features/one-click-pr-or-gist';
 import './features/link-to-prior-blame-line';
 import './features/dim-bots';
 import './features/conflict-marker';
